Move HeaderErrorWidget into a separate file
